+++ examples/data-sources/meraki_appliance_vpn_firewall_rules/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_vpn_firewall_rules" "example" {
+  organization_id = "123456"
+}