+++ docs/guides/changelog.md
@@ -47,6 +47,7 @@ description: |-
 - Add `meraki_appliance_uplink_settings` resource and data source
 - Add `meraki_appliance_site_to_site_vpn` resource and data source
 - Add `meraki_appliance_third_party_vpn_peers` resource and data source
+- Add `meraki_appliance_vpn_firewall_rules` resource and data source
 
 ## 0.1.1